requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
constantly==15.1.0
cryptography==41.0.2
daphne==4.0.0
distlib==0.3.6
Django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
django-pagination==1.0.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
filelock==3.12.2
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
install==1.3.5
Markdown==3.4.4
msgpack==1.0.5
packaging==23.1
Pillow==10.0.0
pipenv==2023.7.11
platformdirs==3.9.0
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
uritemplate==4.1.1
virtualenv==20.24.0
virtualenv-clone==0.5.7
zope.interface==6.0
